More work?? guhhh

+++ Projectiles/Melee/ElementalGlassStar.cs
@@ -80,7 +80,7 @@ namespace CalamityMod.Projectiles.Melee
 				Main.gore[starGore].velocity += projectile.velocity * 0.3f;
 			}
 
-			if (projectile.velocity.Length() < 2f) 
+			if (projectile.velocity.Length() < 2f && projectile.timeLeft < (int)(MaxTime / 2f)) 
 			{
 				projectile.Kill();
 				return;
